+++ tests/test_baker.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 
 import pytest
 from django.conf import settings
+from django.contrib.contenttypes.models import ContentType
 from django.db.models import Manager
 from django.db.models.signals import m2m_changed
 from django.test import TestCase, override_settings
@@ -570,7 +571,7 @@ class TestBakerCreatesAssociatedModels(TestCase):
         assert 0 == models.RelatedNamesModel.objects.count()
 
     def test_ensure_reverse_fk_for_many_to_one_is_working(self):
-        """This is a regression test to make sure issue 291 is fixed"""
+        """This is a regression test to make sure issue 291 is fixed."""
         fk1, fk2 = baker.prepare(
             models.Issue291Model3, fk_model_2=None, name="custom name", _quantity=2
         )
@@ -609,13 +610,23 @@ class TestHandlingContentTypeField:
     def test_create_model_with_contenttype_field(self):
         dummy = baker.make(models.DummyGenericForeignKeyModel)
         assert isinstance(dummy, models.DummyGenericForeignKeyModel)
+        assert isinstance(dummy.content_type, ContentType)
 
 
-@pytest.mark.django_db
 class TestHandlingContentTypeFieldNoQueries:
     def test_create_model_with_contenttype_field(self):
-        dummy = baker.prepare(models.DummyGenericForeignKeyModel)
+        # Clear ContentType's internal cache so that it *will* try to connect to
+        # the database in order to fetch the corresponding ContentType model for
+        # a randomly chosen model.
+        ContentType.objects.clear_cache()
+
+        with pytest.warns(
+            UserWarning,
+            match="Database access disabled, returning ContentType raw instance",
+        ):
+            dummy = baker.prepare(models.DummyGenericForeignKeyModel)
         assert isinstance(dummy, models.DummyGenericForeignKeyModel)
+        assert isinstance(dummy.content_type, ContentType)
 
 
 @pytest.mark.django_db